Added more accurate projectile shots. Bug fix in Clock.java. Towers shoot at the closest target.

+++ src/net/johnbrooks/fjg/Clock.java
@@ -55,7 +55,7 @@ public class Clock
         totalTime += delta;
     }
 
-    public static void changeMultiplier(int change)
+    public static void changeMultiplier(float change)
     {
         if (multiplier + change < -1 || multiplier + change > 7)
         {
